# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample send request use API.",
        "title": "Application EventHub",
        "contact": {},
        "version": "1.0"
    },
    "host": "172.17.0.1:2027",
    "basePath": "/",
    "paths": {
...
        "/communities/{id}": {
            "get": {
                "description": "Retrieve detail information of a community by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Get Detail Community",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/communities/{id}/join": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Join a community based on the provided community ID in path parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Join Community",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/communities/{id}/leave": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "leave a community based on the provided community ID in path parameter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "LeaveCommunity",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/communities/{id}/members": {
            "get": {
                "description": "Retrieve all members associated with a specific community ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Get Community Members",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Community ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "failed"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer Token used as identity for accessing backend",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}